// types/ethers-contracts/StakingBoost.ts
  export interface StakingBoostInterface extends Interface {
    getFunction(nameOrSignature: "BOOST_PER_TOKENS" | "LOCK_PERIOD" | "getBoost" | "getStakeInfo" | "shieldToken" | "stake" | "stakes" | "totalStaked" | "withdraw"): FunctionFragment;

    getEvent(nameOrSignatureOrTopic: "Staked" | "Withdrawn"): EventFragment;

    encodeFunctionData(functionFragment: 'BOOST_PER_TOKENS', values?: undefined): string;
encodeFunctionData(functionFragment: 'LOCK_PERIOD', values?: undefined): string;
encodeFunctionData(functionFragment: 'getBoost', values: [AddressLike]): string;
encodeFunctionData(functionFragment: 'getStakeInfo', values: [AddressLike]): string;
encodeFunctionData(functionFragment: 'shieldToken', values?: undefined): string;
encodeFunctionData(functionFragment: 'stake', values: [BigNumberish]): string;
encodeFunctionData(functionFragment: 'stakes', values: [AddressLike]): string;
encodeFunctionData(functionFragment: 'totalStaked', values?: undefined): string;
encodeFunctionData(functionFragment: 'withdraw', values: [BigNumberish]): string;

    decodeFunctionResult(functionFragment: 'BOOST_PER_TOKENS', data: BytesLike): Result;
decodeFunctionResult(functionFragment: 'LOCK_PERIOD', data: BytesLike): Result;
decodeFunctionResult(functionFragment: 'getBoost', data: BytesLike): Result;
decodeFunctionResult(functionFragment: 'getStakeInfo', data: BytesLike): Result;
decodeFunctionResult(functionFragment: 'shieldToken', data: BytesLike): Result;
decodeFunctionResult(functionFragment: 'stake', data: BytesLike): Result;
decodeFunctionResult(functionFragment: 'stakes', data: BytesLike): Result;
decodeFunctionResult(functionFragment: 'totalStaked', data: BytesLike): Result;
decodeFunctionResult(functionFragment: 'withdraw', data: BytesLike): Result;
  }

  
    export namespace StakedEvent {
      export type InputTuple = [user: AddressLike, amount: BigNumberish];
      export type OutputTuple = [user: string, amount: bigint];
      export interface OutputObject {user: string, amount: bigint };
      export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>
      export type Filter = TypedDeferredTopicFilter<Event>
      export type Log = TypedEventLog<Event>
      export type LogDescription = TypedLogDescription<Event>
    }

  

    export namespace WithdrawnEvent {
      export type InputTuple = [user: AddressLike, amount: BigNumberish];
      export type OutputTuple = [user: string, amount: bigint];
      export interface OutputObject {user: string, amount: bigint };
      export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>
      export type Filter = TypedDeferredTopicFilter<Event>
      export type Log = TypedEventLog<Event>
      export type LogDescription = TypedLogDescription<Event>
    }

  

  export interface StakingBoost extends BaseContract {
    
    connect(runner?: ContractRunner | null): StakingBoost;
    waitForDeployment(): Promise<this>;

    interface: StakingBoostInterface;

    
  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined,
  ): Promise<Array<TypedEventLog<TCEvent>>>
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(event: TCEvent, listener: TypedListener<TCEvent>): Promise<this>
  on<TCEvent extends TypedContractEvent>(filter: TypedDeferredTopicFilter<TCEvent>, listener: TypedListener<TCEvent>): Promise<this>
  
  once<TCEvent extends TypedContractEvent>(event: TCEvent, listener: TypedListener<TCEvent>): Promise<this>
  once<TCEvent extends TypedContractEvent>(filter: TypedDeferredTopicFilter<TCEvent>, listener: TypedListener<TCEvent>): Promise<this>

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>
  removeAllListeners<TCEvent extends TypedContractEvent>(event?: TCEvent): Promise<this>


    
    
    BOOST_PER_TOKENS: TypedContractMethod<
      [],
      [bigint],
      'view'
    >
    

    
    LOCK_PERIOD: TypedContractMethod<
      [],
      [bigint],
      'view'
    >
    

    
    getBoost: TypedContractMethod<
      [user: AddressLike, ],
      [bigint],
      'view'
    >
    

    
    getStakeInfo: TypedContractMethod<
      [user: AddressLike, ],
      [[bigint, bigint, bigint] & {amount: bigint, stakedAt: bigint, unlockTime: bigint }],
      'view'
    >
    

    
    shieldToken: TypedContractMethod<
      [],
      [string],
      'view'
    >
    

    
    stake: TypedContractMethod<
      [amount: BigNumberish, ],
      [void],
      'nonpayable'
    >
    

    
    stakes: TypedContractMethod<
      [arg0: AddressLike, ],
      [[bigint, bigint] & {amount: bigint, stakedAt: bigint }],
      'view'
    >
    

    
    totalStaked: TypedContractMethod<
      [],
      [bigint],
      'view'
    >
    

    
    withdraw: TypedContractMethod<
      [amount: BigNumberish, ],
      [void],
      'nonpayable'
    >
    


    getFunction<T extends ContractMethod = ContractMethod>(key: string | FunctionFragment): T;

    getFunction(nameOrSignature: 'BOOST_PER_TOKENS'): TypedContractMethod<
      [],
      [bigint],
      'view'
    >;
getFunction(nameOrSignature: 'LOCK_PERIOD'): TypedContractMethod<
      [],
      [bigint],
      'view'
    >;
getFunction(nameOrSignature: 'getBoost'): TypedContractMethod<
      [user: AddressLike, ],
      [bigint],
      'view'
    >;
getFunction(nameOrSignature: 'getStakeInfo'): TypedContractMethod<
      [user: AddressLike, ],
      [[bigint, bigint, bigint] & {amount: bigint, stakedAt: bigint, unlockTime: bigint }],
      'view'
    >;
getFunction(nameOrSignature: 'shieldToken'): TypedContractMethod<
      [],
      [string],
      'view'
    >;
getFunction(nameOrSignature: 'stake'): TypedContractMethod<
      [amount: BigNumberish, ],
      [void],
      'nonpayable'
    >;
getFunction(nameOrSignature: 'stakes'): TypedContractMethod<
      [arg0: AddressLike, ],
      [[bigint, bigint] & {amount: bigint, stakedAt: bigint }],
      'view'
    >;
getFunction(nameOrSignature: 'totalStaked'): TypedContractMethod<
      [],
      [bigint],
      'view'
    >;
getFunction(nameOrSignature: 'withdraw'): TypedContractMethod<
      [amount: BigNumberish, ],
      [void],
      'nonpayable'
    >;

    getEvent(key: 'Staked'): TypedContractEvent<StakedEvent.InputTuple, StakedEvent.OutputTuple, StakedEvent.OutputObject>;
getEvent(key: 'Withdrawn'): TypedContractEvent<WithdrawnEvent.InputTuple, WithdrawnEvent.OutputTuple, WithdrawnEvent.OutputObject>;

    filters: {
      
      'Staked(address,uint256)': TypedContractEvent<StakedEvent.InputTuple, StakedEvent.OutputTuple, StakedEvent.OutputObject>;
      Staked: TypedContractEvent<StakedEvent.InputTuple, StakedEvent.OutputTuple, StakedEvent.OutputObject>;
    

      'Withdrawn(address,uint256)': TypedContractEvent<WithdrawnEvent.InputTuple, WithdrawnEvent.OutputTuple, WithdrawnEvent.OutputObject>;
      Withdrawn: TypedContractEvent<WithdrawnEvent.InputTuple, WithdrawnEvent.OutputTuple, WithdrawnEvent.OutputObject>;
    
    };
  }